Revert "Merge branch 'tracker_features' of https://github.com/ProjectGoose/Super-Puper-AR-RPG-Quest into tracker_feature"
