+++ app/src/main/AndroidManifest.xml
@@ -35,8 +35,7 @@
 
         <activity
             android:name=".MapsActivity"
-            android:label="@string/title_activity_maps"
-            android:noHistory="true"/>
+            android:label="@string/title_activity_maps"></activity>
         <activity android:name=".RegistryActivity" />
         <activity android:name=".LoginActivity">
             <intent-filter>
