+++ app/src/main/java/com/example/superpuper_ar_rpg/LoginActivity.java
@@ -8,17 +8,12 @@ import android.os.Bundle;
 import android.view.View;
 import android.view.WindowManager;
 import android.widget.Button;
-import android.widget.CompoundButton;
 import android.widget.EditText;
-import android.widget.RadioGroup;
-import android.widget.Switch;
 
 public class LoginActivity extends AppCompatActivity {
 
     Button btn_login, btn_registry;
     EditText et_email, et_password;
-
-    Switch sw_mocker; //debug
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -31,9 +26,6 @@ public class LoginActivity extends AppCompatActivity {
         btn_registry = findViewById(R.id.btn_register);
         et_email = findViewById(R.id.et_email);
         et_password = findViewById(R.id.et_password);
-
-        sw_mocker = findViewById(R.id.swtch_mocker); //debug
-
 
         if (getIntent() != null) {
             et_email.setText(getIntent().getStringExtra(Intent.EXTRA_EMAIL));
@@ -57,17 +49,6 @@ public class LoginActivity extends AppCompatActivity {
                 Intent goToRegistry = new Intent(context, RegistryActivity.class);
                 goToRegistry.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK);
                 startActivity(goToRegistry);
-            }
-        });
-
-        //debug
-        sw_mocker.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener(){
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                if(isChecked)
-                    MapHandler.isMocking = true;
-                else
-                    MapHandler.isMocking = false;
             }
         });
     }
