+++ app/src/main/java/com/example/superpuper_ar_rpg/MapsActivity.java
@@ -1,5 +1,6 @@
 package com.example.superpuper_ar_rpg;
 
+
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.FragmentActivity;
@@ -9,87 +10,68 @@ import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.pm.ActivityInfo;
 import android.content.pm.PackageManager;
+import android.location.Location;
+import android.location.LocationListener;
 import android.location.LocationManager;
 import android.os.Bundle;
 import android.provider.Settings;
 import android.util.Log;
 import android.view.WindowManager;
+import android.widget.Toast;
 
 import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.MarkerOptions;
 
 
-public class MapsActivity extends FragmentActivity{
+public class MapsActivity extends FragmentActivity implements OnMapReadyCallback {
 
-    private SupportMapFragment mapFragment;
+    private GoogleMap mMap;
+    private FusedLocationProviderClient mFusedLocationClient;
+//    private int locationRequestCode = 1000;
     private LocationManager mLocationManager;
-
-    MapHandler mapHandler;
+    private double wayLatitude =  43, wayLongitude = 169;
+    private SupportMapFragment mapFragment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_maps);
 
-        //Портретная ориентация, чтобы не было лишней головной боли пока
-        setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
-        //А вот тут убираем стандартную верхнюю панель
+        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this);
+        mLocationManager = (LocationManager) getSystemService(LOCATION_SERVICE); //привязываем locationManager к сервису контроля за gps???
+
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
-
-        //привязываем locationManager к сервису контроля за gps???
-        mLocationManager = (LocationManager) this.getSystemService(LOCATION_SERVICE);
-        // Получаем SupportMapFragment и получаем callback когда карта будет готова к работе
+        // Obtain the SupportMapFragment and get notified when the map is ready to be used.
         mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
+        mapFragment.getMapAsync(this);
 
-        mapHandler = new MapHandler(mapFragment, mLocationManager, this);
-        Log.d("ActivityState", "onCreate");
+
     }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-        Log.d("ActivityState", "onStart");
-    }
-
     @Override
     protected void onResume(){
         super.onResume();
-        //проверям, разрешено ли использовать gps (надо дописать)
-        if (ContextCompat.checkSelfPermission(MapsActivity.this, Manifest.permission.ACCESS_FINE_LOCATION)
-                != PackageManager.PERMISSION_GRANTED) {
-            //TODO Permission is not granted
-        }
         //Проверяем, включен ли gps
-        if (!mLocationManager.isProviderEnabled("gps")) {
+        if(!mLocationManager.isProviderEnabled("gps")){
             GPSEnableDialog gpsEnable = new GPSEnableDialog();
             gpsEnable.show(getSupportFragmentManager(), "what`s the tag?");
         }
-            //Если всё включено, то запускаем хэндлер
-            mapHandler.start();
 
-        Log.d("ActivityState", "onResume");
-    }
+        //проверям, разрешено ли использовать gps (надо дописать)
+        if (ContextCompat.checkSelfPermission(MapsActivity.this, Manifest.permission.ACCESS_FINE_LOCATION)
+                != PackageManager.PERMISSION_GRANTED) {
+            // Permission is not granted
+        }
 
-    @Override
-    protected void onPause() {
-        super.onPause();
-        Log.d("ActivityState", "onPause");
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        Log.d("ActivityState", "onStop");
-        mapHandler.stop();
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        Log.d("ActivityState", "onDestroy");
+        //Обновляем координаты вызывая mlocationListener через callback
+        mLocationManager.requestLocationUpdates(mLocationManager.GPS_PROVIDER, 10000, 1, mlocationListener);
     }
 
     //создаем окно с просьбой включить gps
@@ -115,5 +97,64 @@ public class MapsActivity extends FragmentActivity{
             });
             return builder.create();
         }
+    }
+
+    //реализуем интерфейс LocationListener
+    LocationListener mlocationListener = new LocationListener() {
+        @Override
+        public void onLocationChanged(Location location) {
+            //при обновлении локации обновляем карту
+            Toast.makeText(MapsActivity.this, "LocationChanged", Toast.LENGTH_SHORT).show();
+            mapFragment.getMapAsync(MapsActivity.this);
+            Log.d("LocationChanged", "Longtitude " + location.getLongitude() + " Latitude " + location.getLatitude());
+        }
+
+        @Override
+        public void onStatusChanged(String s, int i, Bundle bundle) {
+
+        }
+
+        @Override
+        public void onProviderEnabled(String s) {
+
+        }
+
+        @Override
+        public void onProviderDisabled(String s) {
+
+        }
+    };
+
+
+
+    /**
+     * Manipulates the map once available.
+     * This callback is triggered when the map is ready to be used.
+     * This is where we can add markers or lines, add listeners or move the camera. In this case,
+     * we just add a marker near Sydney, Australia.
+     * If Google Play services is not installed on the device, the user will be prompted to install
+     * it inside the SupportMapFragment. This method will only be triggered once the user has
+     * installed Google Play services and returned to the app.
+     */
+    @Override
+    public void onMapReady(GoogleMap googleMap) {
+        mMap = googleMap;
+
+
+        mFusedLocationClient.getLastLocation().addOnSuccessListener(this, location -> {
+                if (location != null) {
+                    LatLng myCoordinates = new LatLng(location.getLatitude(), location.getLongitude());
+                    Log.d("TAG2", "Latitude " + wayLatitude + " Longtitude " + wayLongitude);
+
+//                  mMap.clear();
+                    mMap.addMarker(new MarkerOptions().position(myCoordinates).title("Marker in ... what is it?"));
+                    mMap.moveCamera(CameraUpdateFactory.newLatLng(myCoordinates));
+                    Toast.makeText(MapsActivity.this, "Ha ha, found you, Lenny", Toast.LENGTH_SHORT).show();
+                }
+                else {
+                    Toast.makeText(MapsActivity.this, "Location == null", Toast.LENGTH_SHORT).show();
+                    Log.e("TAG3", "Latitude " + wayLatitude + " Longtitude " + wayLongitude);
+                }
+        });
     }
 }
