+++ app/src/main/res/layout/activity_login.xml
@@ -66,12 +66,6 @@
             android:layout_height="wrap_content"
             android:minWidth="300dp"
             android:text="@string/form_registry" />
-
-        <Switch
-            android:id="@+id/swtch_mocker"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Mock location?" />
     </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
